Adjusted returned values to unsigned long

+++ fib.c
@@ -2,11 +2,11 @@
 #include <stdlib.h>
 #include <string.h>
 
-int fib_iterative(int n) {
+unsigned long fib_iterative(int n) {
    int i;
-   int next;
-   int num1 = 0;
-   int num2 = 1;
+   unsigned long next;
+   unsigned long num1 = 0;
+   unsigned long num2 = 1;
    if (n <= 1) {
       return 0;
    }
@@ -18,7 +18,7 @@ int fib_iterative(int n) {
    return num2;
 }
 
-int fib_recursive(int n) {
+unsigned long fib_recursive(int n) {
    if (n == 0 || n == 1) {
       return n;
    }
@@ -40,10 +40,10 @@ int main(int argc, char* argv[]) {
       int nth = file_num + atoi(argv[1]);
 
       if (strcmp(argv[2], "-i") || strcmp(argv[2], "i")) {
-         printf("%d", fib_iterative(nth));
+         printf("%lu", fib_iterative(nth));
       }
       else if (strcmp(argv[2], "-r") || strcmp(argv[2], "r")) {
-         printf("%d", fib_recursive(nth - 1));
+         printf("%lu", fib_recursive(nth - 1));
       }
       else {
          printf("Invalid argument.");
